+++ core/utils.lua
@@ -1,12 +1,5 @@
-local validEquipLocs = VALID_INV_TYPE
-
-function extensiveTypeisValid(itemEquipLoc)
-    return validEquipLocs[itemEquipLoc] or false
-end
-
 -- gold format display
 function formatGoldAmount(val)
-    -- Find the comma-separated number
     local commaSepNr = val:match("(%d[%d,]*%d)")
     if not commaSepNr then
         return val
@@ -27,12 +20,34 @@ function formatGoldAmount(val)
     return result
 end
 
-function textCFormat(color,text)
-    return "\124cn"..color..":"..text.."\124r"
+function textCFormat(color, text)
+    return "\124cn" .. color .. ":" .. text .. "\124r"
 end
 
+function getItemDetails(itemInfo)
+    local itemName, itemLink, _, _, _, itemType, _, _, itemEquipLoc, _, _, _, _, bindType = C_Item.GetItemInfo(itemInfo)
+    local result = {
+        itemId = itemInfo,
+        name = itemName,
+        link = itemLink,
+        itemType = itemType,
+        itemEquipLoc = itemEquipLoc,
+        bindType = bindType,
+    }
+    return result
+end
 
--- Constructors
+-- ========================
+-- Section: Constructors
+-- ========================
+
+---@param params CH.CustomFontConstructorOptions
+function fontBuilder(params)
+    local result = params.parent:CreateFontString(nil, "OVERLAY", "GameTooltipText")
+    result:SetPoint(params.point.pos, params.point.x, params.point.y)
+    result:SetText(params.text)
+    return result
+end
 
 ---@param params CH.CustomFrameConstructorOptions
 function frameBuilder(params)
@@ -50,6 +65,14 @@ function frameBuilder(params)
     })
     result:SetBackdropColor(0.05, 0.05, 0.05, 0.95)
     result:SetBackdropBorderColor(0, 0, 0, 1)
+    -- create frame title text
+    if params.titleBuilder ~= nil then
+        fontBuilder({
+            parent = result,
+            text = params.titleBuilder.text,
+            point = params.titleBuilder.point
+        })
+    end
     return result
 end
 
@@ -61,7 +84,15 @@ function buttonBuilder(params)
     result:SetWidth(params.width)
     result:SetHeight(params.height)
     result:SetPoint(params.point.pos, params.point.x, params.point.y)
+
+    -- Skin for ElvUI.
+    pcall(function(result)
+        local E = ElvUI[1]
+        local S = E:GetModule("Skins")
+        if E.private.skins.blizzard.enable and E.private.skins.blizzard.merchant then
+            S:HandleButton(result)
+        end
+    end, result)
+
     return result
 end
-
-